refactor: relocate WebAppHead component
